type cast tried with char to ascci value

+++ input-output.cpp
@@ -1,9 +1,9 @@
 #include<iostream>
 int main()
 {
-long long int x;
-char a;
-std::cin >> x >> a;
-std::cout << x <<std::endl << a;
-return 0;
+    long long int x;
+    char a;
+    std::cin >> x >> a;
+    std::cout << x <<std::endl << (int)a;
+    return 0;
 }
